Create an OAuth Client for LinkedIn API

The li_client function initializes an OAuth client to authenticate and interact with the LinkedIn API. It uses environment variables to retrieve the LinkedIn application credentials (client ID and client secret).

Usage

li_client()

Value

An OAuth client object of class httr2_request that can be used to authenticate and make requests to the LinkedIn API.

Details

The function uses the httr2::oauth_client function to create an OAuth client for LinkedIn. The following environment variables must be set:

LI_CLIENT_ID

The LinkedIn application client ID.

LI_CLIENT_SECRET

The LinkedIn application client secret.

Examples

if (FALSE) { # \dontrun{
# Create an OAuth client for LinkedIn
client <- li_client()

client
} # }